move ImmutableRowsIndex to de.metas.ui.web.view.template
