+++ src/main/java/de/metas/ui/web/bankstatement_reconciliation/PaymentToReconcileRows.java
@@ -10,8 +10,8 @@ import org.compiere.model.I_C_Payment;
 import com.google.common.collect.ImmutableSet;
 
 import de.metas.payment.PaymentId;
-import de.metas.ui.web.payment_allocation.ImmutableRowsIndex;
 import de.metas.ui.web.view.template.IRowsData;
+import de.metas.ui.web.view.template.ImmutableRowsIndex;
 import de.metas.ui.web.window.datatypes.DocumentId;
 import de.metas.ui.web.window.datatypes.DocumentIdsSelection;
 import lombok.Builder;
